Update README.md
o Do: 1 per student in the group!

Create an OS Task Sequence that will deploy Windows Server 2022 as 'Available' to your Group's collection. Once you have created it, show the 'View' showing all the steps within the Task Sequence, and show that it shows up on your devices in Software Center.
Create an OS Task Sequence that will deploy an Upgrade to Windows Server 2022 as 'Available' to your Group's collection
Go through the Task Sequence Media Wizard to create 'Stand-alone media'. Document your steps and end-result.
Go through the Task Sequence Media Wizard to create 'Bootable media'. Document your steps and end-result. NOTE: You will have a step where you have to add a Cert to the Bootable media, the cert is available at "\\asu-sccm\BootMediaCert", along with the password you need for it.
For the purposes of the Lab, you do NOT have to run the deployment of any of these Task Sequences to your Lab VM. In fact, attempting to do so could lead to getting your Lab VM in a non-working state.